Weight lane line likelihood by pixel value
Earlier versions just looked for non-zero pixel values in the
threshold image and treated them all equally.

This commit weights the importance of each pixel in the threshold
image by its value so that more prevalent pixels count more when
deciding on line line inliers and the subsequent best-fit line.

Note that the threshold image does not need to be a binary mask.

**Warning** - current implementation duplicates pixel index based on
pixel value so larger values will require more space and time for
processing.  Try to keep pixel values < 10